Add view classes to the Remixer project
Adds the remixer_ui module which is an android library module. This
contains the classes and interfaces needed to display a UI for Remixer.

- Adds the RemixWidget family of classes. RemixWidget is a generic
class that is implemented by any UI control view for remixes. This
allows to set the Remix on the widget and have it call setValue(…) on
the remix as a callback for user input. They also correspond 1:1 to
layout files.
- Adds the view package which provides a RemixerFragment and code to
display all remixes in an activity.
- in remixer_core adds a RemixerActivity interface to the view package
that is meant to be implemented by any activity that uses remixer. This
will allow the dialogs to interact with the Remixes for the current
activity.