Add support for default layout to inflate for Remixes.
Summary:
Up until now you had to specify a layoutId to use Remixes. While that makes sense when we offer multiple layouts, it is very inconvenient. This adds the ability for the framework to assume you're using a default layout if you don't specify it.

This will be particularly useful when we implement annotation processing since with sensible defaults we can make a method a remix with just one @StringRemixMethod annotation (for example).

The reasoning behind not using default values in the remix classes themselves is that the R class is generated in a separate submodule that depends on the remixer_core submodule, this allows us to use annotation processing in the future (Diff incoming). So the DefaultValues are being set in the project that defines it.

Test Plan: Updated the example app, it still works perfectly with this.

Reviewers: carlsz, O5 Material Remixer, LeFrenchTouch

Reviewed By: O5 Material Remixer, LeFrenchTouch

Tags: #material_remixer

Differential Revision: http://codereview.cc/D1527