Major name refactor: Remix turns into Variable.
Notice the product name still remains as Remixer, but the atomic object, Remix, was renamed to Variable to make it less confusing.
So now, a Remixer object contains * variables and triggers.

I made sure to reflect this name change across all class names, field and variable names, constants, strings, javadocs, everything I could think of. That's why this is a large change. It is a large change, but significantly simple.

I took the chance to remove a couple things that were unnecessary (some font styles, for example), but other than that it's onlyname changes.