Add second activity to example app
MainActivity contains a two textviews, title and freeform text, whose
text is controlled by a String list Variable and a String Variable,
respectively.

Box activity contains a title text and a box. The box has a color
controlled by a Color list Variable and it is hidden or shown based on
a Boolean variable.

Both activities’ title text share a Range Variable that dictates their
size, this is meant to showcase Remixer keeping values across different
contexts, but this depends on #36