Move serialization from remixer_storage to remixer_core (#89)
* Add a type-safe data type to RemixerItem.

This is still incomplete. Annotation-processing is not working because it hasn’t been adapted to use it. I am planning to move to use the Builders first so it’s easier to adapt.

* Make annotation processing use the Builders for every annotation.

This large change cleans up a lot of the annotation processing code, making a lot of it more reusable. It means all of the common fields for a Remixer Item are initialized by `MethodAnnotation` instead of being initialized all over the place. It improves readability and maintainability of this component, and lets us more easily adapt to changes in the API.

In order to do this I had to add a builder for Trigger that lacked one and make all builders inherit from a `RemixerItem.Builder` class.

* Move all tests in remixer_core and remixer_ui to use Builders

* Move all of the serialization logic into `remixer_core`

That allows us to use it directly in `remixer_core` to improve variable value synchronization and build better foundation for cloud sync.

* Fix class<T> and assertions.