Obliterate triggers.
Triggers have been deemed a maintenance hell for a while now. I had this pending a while ago.

So currently Triggers and Variables share a lot of code through the RemixerItem Interface. I am not yet removing that one, will do that on a separate PR to keep this manageable.

This PR removes all mentions of triggers anywhere and all the code to handle them.

Fixes #99