Decouple the type we use in runtime to represent values from the serialization.
In most cases both are the same, but colors are their special little snowflakes, because they have a very different representation across platforms and the best we can do is give them their own unambiguous representation that doesn’t map directly to Android’s.

For context: Android uses 32-bit integers, encoded as AARRGGBB, web uses either four 0-255 integers representing each component or one #RRGGBBAA string (notice the different order of the components), and iOS has a native object that can be initialized with four 0-255 integers, only. This is why we decided to represent it as an object that has four 0-255 integers, a, r, g and b.

This already works when serializing to JSON but it doesn’t when serializing to firebase, where it expects a more natural java object representation, for this purpose I wrote a new java class that represents this structure, and refactored our serialization code to use it.