Remove dependency on the remixer project which breaks the proguard step.
We remove it on the remixer_example project by switching them to the explicit dependencies and making the remixer_annotation dependency a `provided` instead of `compile`.

The problem with `compile` dependencies is that the classes in remixer_annotation are added to the classpath. Those classes depend on the annotation processing classes (javax.model.*, javax.annotation.processing.* ...) in the JDK which are unavailable in the Android SDK, so when proguard runs it cannot resolve those classes and fails.

A lot of those errors are visible on a TravisCI run explicitly designed to make it fail: https://travis-ci.org/material-foundation/material-remixer-android/builds/200502951