Remove the demo UI and replace it with a more real-ish looking demo.
This demo has a randomly generated list of transactions a stereotypical Software Engineer could be making on his credit card. There is a FAB that is used to trigger the RemixerFragment. You can change the header color, whether icons are displayed or not, the size of the total transaction in period number, and whether we use last month vs this month to calculate.